+++ CHANGELOG.md
@@ -14,8 +14,9 @@ All notable changes to this project will be documented in this file.
 - Add support for recording gRPC stats.
 - Add support for recording HTTP stats.
 - Enforce `--strictNullChecks` and `--noUnusedLocals` Compiler Options on [opencensus-instrumentation-http], [opencensus-instrumentation-grpc] and [opencensus-propagation-tracecontext] packages.
+- Enforce `--strictNullChecks` and `--noUnusedLocals` Compiler Options on  [opencensus-exporter-zipkin] packages.
 
-## 0.0.9 - 2019-02-12
+## 0.0.9 - 2019-02-12p
 - Add Metrics API.
 - Add Resource API.
 - Add Tags API.